extract to initial screen
